api/requirements.txt
aiobcrypt==3.2.0.post1
aiofile==3.8.8
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==3.2.2
caio==0.9.13
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
databases==0.8.0
dill==0.3.7
distlib==0.3.7
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
miniopy-async==1.16
multidict==6.0.4
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
psycopg2==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic-async-validation==0.1.0
pydantic_core==2.10.1
Pygments==2.11.2
pyhumps==3.8.0
PyJWT==2.8.0
pylint==3.0.1
pytest==7.4.2
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
virtualenv==20.24.5
XlsxWriter==3.1.9
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
